updated unit tests to pass
